requirements.txt
Django>=5.0,<6.0
gunicorn>=21.2
psycopg2-binary>=2.9
dj-database-url>=2.2
python-dotenv>=1.0
whitenoise>=6.6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.3
drf-spectacular>=0.27